GNSS change only: support use of UART 2.
In recent FW versions the ZED-F9P GNSS device supports reception of messages over either of its two UARTs, hence with this commit we support use of either one.

Note that most of the ubxlib code doesn't care about which UART port of the GNSS device it is connected to; it is only needed for cases where the output protocol of the GNSS device has to be set (which is port-specific) and when using commands which have no acknowledgement (in which case the ubxlib code asks the GNSS device how many messages it has received on a given port to determine whether message reception was successful).  In other words, pretty much everywhere U_GNSS_TRANSPORT_UART_2 == U_GNSS_TRANSPORT_UART_1 == U_GNSS_TRANSPORT_UART.